Move profile related view to own module
